Renamed abin commands
