+++ src/hatchet/__main__.py
@@ -9,14 +9,13 @@ import warnings
 import hatchet
 
 from hatchet.utils.preprocess import main as preprocess
-from hatchet.utils.countPos import main as countPos
-from hatchet.utils.cluBB_KDE import main as kdeBB
-from hatchet.utils.adaptiveBin import main as abin
-from hatchet.utils.formArray import main as array
+from hatchet.utils.cluster_bins_kde import main as cluster_bins_kde
 from hatchet.utils.count_reads import main as count_reads
+from hatchet.utils.count_reads import main as count_reads_fw
 from hatchet.utils.genotype_snps import main as genotype_snps
 from hatchet.utils.count_alleles import main as count_alleles
 from hatchet.utils.combine_counts import main as combine_counts
+from hatchet.utils.combine_counts import main as combine_counts_fw
 from hatchet.utils.cluster_bins import main as cluster_bins
 from hatchet.utils.plot_bins import main as plot_bins
 from hatchet.bin.HATCHet import main as compute_cn
@@ -27,8 +26,9 @@ from hatchet.utils.Phase import main as Phase
 
 solve_bin = os.path.join(os.path.dirname(hatchet.__file__), 'solve')
 
-commands = ('count-reads', 'genotype-snps', 'count-alleles', 'combine-counts', 'cluster-bins', 'plot-bins',
-            'compute-cn', 'plot-cn', 'run', 'preprocess', 'countPos', 'kdeBB', 'abin', 'array', 'PhasePrep', 'Phase')
+commands = ('count-reads', 'count-reads-fw', 'genotype-snps', 'count-alleles', 'combine-counts',
+            'combine-counts-fw', 'cluster-bins', 'plot-bins','compute-cn', 'plot-cn', 'run',
+            'PhasePrep', 'Phase', 'cluster-bins-kde')
 
 def print_usage():
     print('HATCHet v' + hatchet.__version__)
@@ -39,10 +39,10 @@ def print_usage():
 def main():
     # support for old command names as they've been used in earlier versions of HATCHet
     aliases = {
-        'binBAM': 'count-reads',
+        'binBAM': 'count-reads-fw',
         'SNPCaller': 'genotype-snps',
         'deBAF': 'count-alleles',
-        'comBBo': 'combine-counts',
+        'comBBo': 'combine-counts-fw',
         'cluBB': 'cluster-bins',
         'BBot': 'plot-bins',
         'solve': 'compute-cn',
